requirements.txt
pydantic>=2.0.0
# chromadb>=0.4.0
python-dotenv>=1.0.0
openai>=1.0.0
anthropic>=0.15.0
groq>=0.5.0
# sentence-transformers>=2.0.0
tiktoken>=0.5.0
lancedb
fastapi
uvicorn[standard]
requests 
beautifulsoup4
duckduckgo-search
langgraph
instructor
crewai>=0.28.8,<0.29.0
crewai_tools